Refactor QueryMethods
