+++ lib/stretchy/relations/query_methods/ids.rb
@@ -1,0 +1,18 @@
+module Stretchy
+  module Relations
+    module QueryMethods
+      module Ids
+        def ids(*args)
+          spawn.ids!(*args)
+        end
+
+        def ids!(*args) # :nodoc:
+          self.ids_values += args
+          self
+        end
+
+        QueryMethods.register!(:ids)
+      end
+    end
+  end
+end
